pyproject.toml
[project]
name = "multi-ai-agent-projects"
version = "0.1.0"
description = ""
authors = [
    {name = "Norbert Lupamo",email = "91078038+lupamo3@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain-groq (>=1.1.0,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "fastapi (>=0.122.0,<0.123.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "streamlit (>=1.51.0,<2.0.0)",
    "langgraph (>=1.0.4,<2.0.0)",
    "langchain-core (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
